+++ CHANGELOG.md
@@ -7,7 +7,19 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [0.3.2] - 2022-10-23
+## [0.4.0] - 2022-10-20
+
+### Added
+
+- Add optional semantic versioning rule [#22](https://github.com/nbbrd/heylogs/issues/22)
+- Add command to scan content of changelog [#26](https://github.com/nbbrd/heylogs/issues/26)
+- Add missing descriptions in CLI [#29](https://github.com/nbbrd/heylogs/issues/29)
+
+### Changed
+
+- Simplify command parameters [#11](https://github.com/nbbrd/heylogs/issues/11)
+
+## [0.3.2] - 2022-10-03
 
 ### Removed
 
@@ -44,7 +56,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/nbbrd/heylogs/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/nbbrd/heylogs/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/nbbrd/heylogs/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/nbbrd/heylogs/compare/v0.2.0...v0.3.0